requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-ckeditor==5.6.1
django-filer==1.4.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-linguist==0.4.0
django-modeltranslation==0.13
django-mptt==0.9.1
django-polymorphic==2.0.3
django-rosetta==0.9.2
django-translations==1.1.1
django-widget-tweaks==1.4.3
easy-thumbnails==2.6
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
Pillow==5.4.1
polib==1.1.0
psycopg2==2.7.7
python-dateutil==1.5
pytz==2018.9
requests==2.21.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.1
whitenoise==4.1.2
